+++ LinuxIsland/src/services/mod.rs
@@ -11,6 +11,7 @@ pub mod battery;
 pub mod clipboard;
 pub mod devices;
 pub mod gmail;
+pub mod hud;
 pub mod mpris;
 pub mod notifications;
 pub mod screenshots;